Using `wp_viewport` protocol for fractional scale.
This commit breaks ABIs.

+++ include/swingby/surface.h
@@ -46,10 +46,10 @@ SB_EXPORT
 void sb_surface_update(sb_surface_t *surface);
 
 SB_EXPORT
-uint32_t sb_surface_scale(const sb_surface_t *surface);
+float sb_surface_scale(const sb_surface_t *surface);
 
 SB_EXPORT
-void sb_surface_set_scale(sb_surface_t *surface, uint32_t scale);
+void sb_surface_set_scale(sb_surface_t *surface, float scale);
 
 /// \memberof sb_surface_t
 /// \brief Set the valid input geometry of the surface.
