+++ src/skia/draw.cpp
@@ -412,7 +412,7 @@ void sb_skia_draw_image3(sb_skia_renderer_t *renderer,
 
 void sb_skia_draw_glyphs(sb_skia_renderer_t *renderer,
                          const sb_rect_t *rect,
-                         uint32_t scale,
+                         float scale,
                          const sb_glyph_layout_t *layout)
 {
     SkCanvas *canvas = _get_canvas(renderer);
